Fix incorrect handling of scapes in redirect URLs
Some mods that previously wouldn't be downloaded correctly due to containing spaces in their filenames will now be requested properly from the server.
This required removing the escaping of the percentage sign ('%') from the redirect strategy, because the server would already answer with an escaped URL. In the future, a more robust solution (i.e. escape detection) might be appropriate.